pyproject.toml
[project]
name = "2025-3-project-05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.86",
    "dotenv>=0.9.9",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pysam>=0.23.3",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "flask>=3.0.0",
    "flask-cors>=5.0.0",
]
